fix ci

+++ src/Voting.sol
@@ -5,29 +5,28 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 contract Voting is Ownable {
     struct Candidate {
         string name;
-        uint voteCount;
+        uint256 voteCount;
     }
 
     Candidate[] public candidates;
-    uint public candidatesCount;
-    uint public votingDeadline;
+    uint256 public candidatesCount;
+    uint256 public votingDeadline;
 
     mapping(address => bool) public hasVoted;
     mapping(address => bool) public isRegistered;
 
     event CandidateAdded(string name);
-    event Voted(address indexed voter, uint indexed candidateId);
+    event Voted(address indexed voter, uint256 indexed candidateId);
     event Registered(address indexed voter);
 
-    constructor(uint _votingDuration) Ownable(msg.sender) {
+    constructor(uint256 _votingDuration) Ownable(msg.sender) {
         votingDeadline = block.timestamp + _votingDuration;
     }
 
     function addCandidate(string memory _name) public onlyOwner {
         for (uint256 i = 0; i < candidates.length; i++) {
             require(
-                keccak256(abi.encodePacked(candidates[i].name)) !=
-                    keccak256(abi.encodePacked(_name)),
+                keccak256(abi.encodePacked(candidates[i].name)) != keccak256(abi.encodePacked(_name)),
                 "Candidate name already exists"
             );
         }
@@ -43,7 +42,7 @@ contract Voting is Ownable {
         emit Registered(msg.sender);
     }
 
-    function vote(uint _candidateId) public {
+    function vote(uint256 _candidateId) public {
         require(block.timestamp < votingDeadline, "Voting has ended");
         require(isRegistered[msg.sender], "You are not registered to vote");
         require(!hasVoted[msg.sender], "You have already voted");
@@ -59,7 +58,7 @@ contract Voting is Ownable {
         return candidates;
     }
 
-    function getVoteCount(uint _candidateId) public view returns (uint) {
+    function getVoteCount(uint256 _candidateId) public view returns (uint256) {
         require(_candidateId < candidates.length, "Invalid candidate");
         return candidates[_candidateId].voteCount;
     }
